+++ api/src/routes/schedule.routes.ts
@@ -1,6 +1,9 @@
 import express from 'express';
 import { ScheduleController } from '../controllers/schedule.controller';
 import { isAdmin, getName } from '../helpers/auth';
+import { ISchedule } from '../interfaces/Schedule';
+const Schedule = require('../models/schedule.model');
+
 
 const router = express.Router();
 const scheduleController: ScheduleController = new ScheduleController();
@@ -113,7 +116,49 @@ const validate_trigger = async (req, res) => {
     } catch (err) {
       res.status(401).send({ message: err });
     }
-  };
-  router.get('/validate_trigger', validate_trigger);
+};
+router.get('/validate_trigger', validate_trigger);
+ 
+  /* Admin: update the entire schedule
+ * @param {*} req
+ * @param {*} res
+ * @return {*} 
+*/
+const update = async (req, res) => {
+    if(!req.headers.authorization) 
+    {
+        res.status(401).send({ message: "This endpoint requires authorization header."});
+        return;
+    }
+    const authToken = req.headers.authorization;
+    const isAdm = await isAdmin(authToken);
+    if(!isAdm)
+    {
+        res.status(401).send({message: "Need admin access."});
+        return;
+    }
+    try
+    {
+        console.log("before json parse\n");
+        // let schedules;
+        let schedules = {} as ISchedule[];
+        // let schedules : ISchedule[4] = {};
+        const numSchedules = req.body.length;
+        for(let i = 0; i < numSchedules; i++)
+        {
+
+            schedules[i] = <ISchedule>req.body[i];
+
+        }
+        await scheduleController.update(schedules, numSchedules);
+        res.status(200).send({message: "Updated schedule."});
+    } catch (err)
+    {
+        res.status(401).send({message: "Error creating schedule: " + err});
+    }
+}
+router.get('/update', update);
+
+
 
 module.exports = router;
